Fix flaky testEnableRollbackTimeoutFailsRollback test
We no longer block on the handler thread in getAvailableRollbacks. That
means it is possible we see an available rollback left over from a
previous test run. Update the test to ensure previous rollbacks are
removed and give sufficient time to make a new rollback available before
concluding that no rollback will be made available in this case.

Test: atest RollbackTest:com.android.tests.rollback.RollbackTest#testEnableRollbackTimeoutFailsRollback
Change-Id: Id94d5452836b07d6f802758bd0dfdd3c276a3c05

Former-commit-id: b61fac95f5c1df1af9efbab9a13ce5d2dc6e80d3