Add new onSmsStatusReportReceived() without message reference
IMS providers must generate unique token and don't need to parse the PDU
to get the message reference for new onSmsStatusReportReceived() API.
The platform will extract the message reference from the PDU and return
it to the IMS providers via acknowledgeSmsReport(). And @Deprecated tag
is added on existing onSmsStatusReportReceived() API.
This CL includes some fixing doc bug and description for APIs also.

Test: atest FrameworksTelephonyTests:ImsSmsDispatcherTest

Change-Id: Ib788e291f2e70c3e062319e17e2c9efe33f31164
Signed-off-by: Taesu Lee <taesu82.lee@samsung.com>

Former-commit-id: 61e826d5763cd6fb88d450f731ddfcc32b6a8aed