Fix bootloop issue with UsageStatsService.
When the configurations are being read from the proto file, if a
duplicate locale is found in the list, an exception is thrown from
LocaleList which causes the UsageStatsService to crash. Instead, we'll
now log a WTF if duplicate locales are found when parsing the
configuration proto and not add the duplicate locale to the list.

Bug: 140197723
Test: atest UsageStatsDatabaseTest
Change-Id: Ie178b3cfdd45e6771a7c0e4bb42e47f79b5c30cc
(cherry picked from commit daa7d729740704171cb5b326441797c979d0dfa3 [formerly 5d9031dbee2fad9b29e6477f5d7af8b622f2b456])

Former-commit-id: 669d886cdc8b577f318152712c6b3b10c8a82599