Prevent animating app from setting exclusion rects
Back gesture still shows if user attempts to swipe
back while we're animating an app away to home screen.
Full screen exclusion rects for launcher are only
granted after the animation has completed, but the
user was able to interact w/ ongoing animations to
see the back arrow.

fixes: 138622418
Test: Open app, swipe home and quickly swipe from
the edge of the screen. Should go to -1 screen and
not see back arrow.

Change-Id: I18603159fbb6b76e0fdb7911cbe7a6e2386c9f87

Former-commit-id: 8009e920c8310d46670c17c09a23941a3522267b