Change commit of pending backup and restore to work per-Rollback.
This changes the pendign backup and restore mechanism in AppDataRollbackHelper
to work on a per-Rollback basis.

This corrects a potential issue where, in the case where two rollbacks
exist for the same package, a restore for one could be cancelled on the
basis of a restore of the other.

This change will also allow per-rollback synchronisation to be added in
future CLs.

Bug: 136241838
Test: atest AppDataRollbackHelperTest
Test: atest RollbackTest
Change-Id: I226fc2aba66b96479f63284d4ba054de64f7fd1c

Former-commit-id: 31f453f30fdd26b4c5d95a3825e38ed3c5580274