Notify AutofillManager that activity is finishing on onDestroy().
It will show the autofill save dialog when the activity is no longer
visible to the user and really to be finished. In current design, it
checks if activity is finishing on onStop(). When activity finishes
through CLEAR_TOP/SINGLE_TOP, it will not trigger autofill save dialog.
So moving the logic to onDestroy().

Bug: 137796463
Test: atest CtsAutoFillServiceTestCases
Test: manual verification with DebugService
Change-Id: Ic751785cfc6d436d74f2ae0f8a0f8292bf68f116

Former-commit-id: c054c5d4a0bc3ae45830ddfbb3d0af8a453a0f5b