Add per-Rollback synchronisation.
This adds a lock object to each Rollback that should be acquired before
accessing any rollback state.

For now, this synchronisation has been added 'in-place' in
RollbackManagerServiceImpl. Future changes will move rollback functionality
from that class into the Rollback class, which will allow the synchronisation
to be handled internally within the Rollback.

Bug: 136241838
Test: atest RollbackTest
Test: atest CtsStagedInstallHostTestCases
Test: atest CtsRollbackManagerHostTestCases
Change-Id: I995f1647ccf0410a8832f097236482552729f4c8

Former-commit-id: 8c0084d980961b0eb3845e0a9a85595f74ad5a72