Handle "uninstall updates" in app-ops-service
When "uninstalling the updates" of a system app the package manager
triggers a ACTION_UID_REMOVED with EXTRA_REPLACING=true. In this case we
want to reset the app-op state, but the app has to still work
afterwards as the UID is not actually removed.

Removing all the state for the UID is not the right behavior
as this does not trigger a re-evualtion of app-op state for this app.

Test: Uninstalled updates for pre-installed Google Photos app and saw
storage related app-ops to be set correctly afterwards
Bug: 140731248

Change-Id: I235dbab877ce4377232799d5d0555ff7d2722ebb

Former-commit-id: 04c5f38246bef2d6b55b2f5ff66f5fc5544ea549