CameraManager: Add @TestApi method getCameraIdListNoLazy().
Add getCameraIdListNoLazy() - a test method which necessarily makes a
call to cameraserver in order to get an updated list of camera ids. This
is needed since camera cts tests use SYSTEM_CAMERA permissions in a way
which are not permitted to happen in real world scenarios : processes
cannot gain and lose SYSTEM_CAMERA permissions for permission checking
purposes (through UiAutomation adopt(drop)ShellPermissionIdentity(). In
that case cameraserver doesn't send callback to the client app informing
it that system cameras aren't available.

Bug: 138130599

Test: camera CTS

Change-Id: Ic00951828ea289a1a04b1d655506ef6474336b91
Signed-off-by: Jayant Chowdhary <jchowdhary@google.com>

Former-commit-id: 556815d71c4c597d2c117e333525dd072d0093d8