Add logging on invocation
Logs when an invocation is started, cancelled, or completed (along with
the invocation type). Also reduces spurious "cancellations" by ignoring
cancellations while we are completing an animation. (This occurs
because Launcher sends progress values back down to zero on TOUCH_UP,
sometimes even if there was a fling, and we don't want to log a
cancelled drag if it was actually completed as a fling.)

Bug: 141195388
Test: manual
Change-Id: I25c04f1c213515def0136c03b89daf331e43d522

Former-commit-id: 35604f0244ad87f5d2763f6f208a8ef3355c88af