[SimpleMonthView] days should be set clickable even when disabled
Currently, ACTION_CLICK is removed when a day is disabled. This causes
the month to be the first non-title talkback focusable item and talkback
places the initial focus on the month and announces all the disabled
dates. According to discussion, disabling an item should not change
whether it is talkback focusable. We still set the days as clickable
even when they are disabled. After the fix, the disabled days are also focusable (and will be
announced as disabled when talkback focus goes there). The initial focus
will go the first non-title focusable item (previous or next month
button).

Bug: 122831096
Test: Tested that after the fix, the initial focus will go the first
non-title focusable item (previous or next month button depends on the
visibility). The month never gets accessibility focus.

Change-Id: I925539282729d5d8418233fc95d3761ea2d0108f

Former-commit-id: a26a37c8ffc9ac8223c26864e35863799da75c11