Replace getPhoneCount with getMaxPhoneCount upon object allocation.
As first step for smooth single SIM to DSDS switch, for DSDS capable
deviced we always allocate objects as if it's in DSDS mode. For example
there will be two Phone objects.
Later we'll evaluate to make the allocations dynamic to save memory.

Bug: 141388730
Test: unittest and manual
Change-Id: I3064eb616371f60776a2930c113582562d206123
Merged-In: I3064eb616371f60776a2930c113582562d206123

Former-commit-id: f87292fa523bf59e32331c4333aa01d9f85c69d1