Move enable/snapshot/restore/delete methods to the Rollback class.
This moves functionality around snapshotting, restoring, and deleting app data,
and enabling rollbacks, into the Rollback class.

Note that the new methods still rely on the AppDataRollbackHelper class - that
functionality may be combined into Rollback in future but those changes are not
in scope for this CL.

This is part of a series of changes to move rollback-specific functionality
into the Rollback class. Once these are completed, synchronisation can be
moved to be internal to Rollback, and the internal state of the Rollback class
can be cleaned up.

Bug: 136241838
Test: atest RollbackTest
Test: atest StagedRollbackTest
Test: atest CtsRollbackManagerHostTestCases
Test: atest RollbackUnitTest (new unit tests added)
Change-Id: I0d3996f1718893046d82e7a0644ea757cc76a01e

Former-commit-id: 856391f52456e4d020bdf5ed394b541fb89f4041