Fix an issue that dumpsys textclassification does not print latest settings
Issue:
When we run dumpsys textclassification, shell spawns a process to run
the dumpsys executable which binds to TCMS and calls TCMS.dump().
In TCMS.dump(), we create a TCM instance, and the TCM instance setups
an ContentObserver to observe the settings changes. However, it does not
pass the checkContentProviderAccess check because AMS does not keep
track of processes spawned from shell.

Solution:
Create the TCM instance using the system server identity.

Fixes: 141690465

Test: Execute the following commands
1. dumpsys textclassification
2. adb shell cmd device_config put textclassifier {key} {value}
3. dumpsys textclassification
4. Verify the output now reflects the latest value.

Change-Id: Ia1185bd02688c0e7a3eb83487d3b843250b7f2cf

Former-commit-id: 384b8a69f9897ac7d3407a6da1fb16e2b9f096f8