Restrict accessible move when tiles cannot be removed
When there were the minimum number of tiles in QSCustomize clicking on a
current tile while on TB would enter ADD MODE, that allowed to move the
tile out.

This CL makes sure that in that case we enter MOVE MODE so the tile
cannot be removed.

Also, makes sure that the tile is focused when entering move mode

Fixes: 140363294
Test: manual with and without TB with 6 and more tiles
Change-Id: I3a27ca1bc54971f6a72b7d69abe6bdb73ca7c56a

Former-commit-id: d06a8494ff8f3b6e1742748a824b41aa61000377