Move Rollback synchronisation into the Rollback class.
This CL moves the last few accesses of Rollback state out of
RollbackManagerServiceImpl and into the Rollback class, and moves
synchronisation for rollbacks into that class. This follows on from
ag/9413749, ag/9431132 and ag/9452004 which moved most of the access
of rollback state previously.

As this change meant that the state of NewRollbacks would no longer be
synchronised, this CL also adds simple per-NewRollback synchronisation.
This may be removed in future changes if the NewRollback class is merged
into the Rollback class.

Bug: 136241838
Test: RollbackTest
Test: StagedRollbackTest
Test: RollbackUnitTest
Test: CtsRollbackManagerHostTestCases
Change-Id: I7a5cf8b5d2bdb090a5a062ac4baa07e4d9e54cda

Former-commit-id: 4af2f75066eab3e9ccb4e8bcaa1560c09377a534