Do not allow input connections from untrusted virtual displays
As Input-routing or Input Method Editor (IME) supported on multi-displays from Q,
user can interact multi-displays with touch and key input.

It could be necessary to consider and prevent a use case that the malicious app
may create in-visible virtual display to steal the window focus or intercept key
or keyboard input through input channel or input connection.

Don't provide display focus to the newly created virtual displays, until the user
interacts with a virtual display, then we give it & receive the display focus.

Bug: 129963294
Test: make sure all MultiDisplay and Window focus related tests pass.
Change-Id: I9bc37695cf6e2e894bf545a69c6cb006666c5539

Former-commit-id: d0b5c8ff80e648506e0c22d9cd4e237b34a21197