Add unit tests for RollbackStore.
This adds a test covering the saving and loading of rollback data as
JSON performed by the RollbackStore class. A small change to RollbackStore
is made to allow for easier testing.

This will allow future changes to the structure of Rollback and to
RollbackStore to be made with greater confidence that the persistence
format has not been changed inadvertently.

Bug: 141915404
Test: atest RollbackStoreTest (new test)
Change-Id: I97b9428f66d0c7ec2f1fd1c7a34afbb4d13bc896

Former-commit-id: d9c78463d92cc1872a00e44ede91f15e9ed4355e