Allow incremental replacement of filegroups with stubs source
Adds a reference to the core-current-stubs-source module (which creates
a srcjar which will contain the public SDK stubs source from all the
modules that make up the core libraries, i.e. the source for
core_current.

This will allow incremental changes to be made in the core library
repositories to gradually switch from using the implementation source
to stubs source to generate the Android SDK public API and JavaDoc
without requiring further changes to frameworks/base.

Once all the required changes have been made then the
core_public_api_files filegroup and all its usages will be removed.

The reason why this will be done incrementally  because bug 142113901
requires some workarounds for some of the modules and it will be easier
to manage if they are done separately.

Bug: 142113521
Test: m -j60 android_stubs_current offline-sdk-docs
Change-Id: I828fda4d64fda789e06296c1d2d5c62e63ee7967

Former-commit-id: e86cc6f022fcad70d26b6a66f95013403656ed60