Dispatch mouse events to the view under cursor.
Events from touchpads may have multiple fingers on them, and the down
event may be active with a finger which coordinate isn't in the view
where the mouse cursor is. Need to make sure it's dispatched to the view
under the mouse cursor.

Bug: 140312476
Test: Two-finger scroll on DualShock 4 always happen in the view under
the cursor, regardless of the order of finger put on the touch pad.
Test: atest FrameworksCoreTests:ViewGroupTest

Change-Id: I679a0a38c830bd418bb412bc6f0d78ba5c93224a

Former-commit-id: 63e6a35508833047c760dceedc05228189a431dc