Replace client token with a new token in InputWindowHandle
The tokens are used to map input data to its window and its channel. Currently InputWindowHandle
tokens uses the client IWindow binder token if the input is associated with a client window or a
newly generated token for temporary surfaces (drag input surface), and windowless surfaces.

These tokens are also used to send policy related events. From the token, it is not obvious what
the token represents for which component. This is a cleanup cl that removes the use of client before adding an explicit token for handling policy related events.
One use case of client token was to support pointer capture but this will move to using
the input channel socket.

Removing the client token allows windows to share its input token with other processes to enable
features like transfer focus without leaking the client token.

Bug: 134365580
Test: go/wm-smoke
Test: atest PointerCaptureTest WindowFocusTests

Change-Id: I49e4c54d977f98e855af9e7ed54443588fdb66c9

Former-commit-id: 18782165207392ce84e478a7554ae958c6d7d841