Add InstallSource to hold trustable installer information.
Create a new class which will hold persisted information about how an
app was installed that cannot be modified, even by the installer of
record or an app with INSTALL_PACKAGES permission.

Add a first field to hold the package name of the app that initiated
the install session.

Plumb this into SessionParams, PackageInstallerSession, InstallParams,
InstallArgs, PackageSettings, PackageProto, ActiveInstallSession etc.

Various minor refactorings along the way.

Bug: 134746019
Test: atest -p services/core/java/com/android/server/pm
Change-Id: I96d9b5034bc98af8c63088df2341e6d4a1fb1b6e

Former-commit-id: 69d2abf4c5a460b6c2b899b0d0455d7d7bd33d59