Use KNOWN_PACKAGES when shared lib consumers
This change ensures we find ALL known packages that could be consuming a
shared library, not only currently installed ones. Without this check,
the system may get into a state in which we have currently uninstalled
but on-device apps that depend on a shared library that does not exist
on device.

Bug: 142083996
Test: Manual; attempt to remove shared lib after marking its consumer uninstalled.
Change-Id: Icf787e1414b85e8d40e9c42131bfeea0a57ecaf3

Former-commit-id: 2e2259831d20feac332447ce68449e6d90297a0c