Use target activity as recents animation target directly.
When recents activity just finished and someone requests to
startRecentsAnimation, although we will launch recents activity, but
the window is not ready when RecentsAnimationController#initialize.
Without set mTargetActivityRecord, we cannot ensure when to skip to set
mLaunchTaskBehind to false in WMS#onAppTransitionFinishedLocked.
We can set the target activity to RecentsAnimationController directly
after AppWindowToken merged into ActivityRecord now.

Fix: 142351430
Test: manually run StartLauncherViaGestureTests
Test: atest WmTests:RecentsAnimationTest
Test: atest WmTests:RecentsAnimationControllerTest

Change-Id: Ib60fb82deb14406ddc60e5897d2baa3d1d2dac75

Former-commit-id: 417b70c1cd29032337b351cc2768a4caa9d20adb