Remove whitelisting REMOTE_BUGREPORT_FINISHED broadcast
After migration to bugreport API, communication between Shell and dumpstate
is done using callbacks and dumpstate does not send this broadcast anymore.

REMOTE_BUGREPORT_FINISHED broadcast was whitelisted previously,
so that remote bugreports are allowed even if DISALLOW_DEBUGGING_FEATURES
user restriction is enforced.

ag/866973 This was necessary so that dumpstate can communicate with
Shell via the broadcast which is not required now.

Bug: 136066578
Bug: 139415553

Test: * Build and flash
      * Install TestDPC
      * Make it device owner
      * Take a remote bugreport. Works as expected
      * Go to "Set user restrictions" and enable "Disallow debugging
        feature"
      * Take a remote bugreport. Works as expected.

Change-Id: I027da8e13ff6f930ec8c33c3aafbb93f0aaf2e9c

Former-commit-id: 8481b9c11b563ab493018e566b524bdc139e6107