Stop task reparenting when enter/exit PIP.
With only one activity in the PIP task, we can simply change the
windowing mode (this was already done implicitly when re-parenting) and
let the display know the existing stack is the new pinned stack as
opposed to creating a new stack just to move the task over.

With multiple activities, we will remain moving the PIP activity into a
separate stack, but we will move it to a stack with the same windowing
mode as the original rather than a pinned one. Then, we will do a
windowing mode change on it also, so it performs the same as the single
activity case.

When exiting, we then simply do a windowing mode change.

Bug: 143772438
Test: Try PIP with one activity (e.g. Youtube) and multiple activities
(VLC), enter/exits works fine. atest PinnedStackTests
Change-Id: Ie961e120762e5e244f14b6ac181509656b107969

Former-commit-id: 6db8fb274e2e81038f569681d7e4ac602f153b4b