Set "today" and "selected" for Datepicker when using Calender
Set the stateDescription to "today" if the date is today. Set selected
if the date is selected. Note that we didn't remove any information, so
this change doesn't need to wait until talkback changes get in.

Bug: 142361188, 141454158
Test: tested with date picker in calender mode. When the focus goes to
today, it announces "today, Oct. 22". When the focus goes to selected
date, it announces "selected, Oct. 22". When it is both today and
selected, it announces "select, today, Oct. 22".

Change-Id: Id210e22ec5ae2bc070c260c9632a605cf2a6236f

Former-commit-id: 7eef9a1adfa002b7ca9721bf6b7e603479635ac8