Support privileged apps installed in APEX.
If an APEX contains a priv-app/ directory, and the APEX package itself
is located in a partition for which privileged apps are allowed, then
the APKs under said directory will be included in the initial scan and
made available to the system.

Bug: 138429615
Test: Move PermissionController to com.android.permission APEX. Run
CtsPermissionTestCases and verify that the 16 failures out of 264 tests,
which seem unrelated to PermissionController, happen also on an
unpatched build on head.
Exempt-From-Owner-Approval: Approved in
https://googleplex-android-review.git.corp.google.com/c/platform/frameworks/base/+/9593749
Merged-In: If809db5d0a061c1a55aeb7f830c183fa822adca5
Change-Id: If809db5d0a061c1a55aeb7f830c183fa822adca5

Former-commit-id: 787471af481e875c59a7720f51174c893f85c97f