Move DozeServiceHost out of StatusBar.
This is an imperfect extraction. That is to say, DozeServiceHost is
successfully moved out of StatusBar, but it is apparent, upon
inspection, that there is not good separation of concerns between
the two classes. There is some functionality that should be fully
moved from on class to to the other. I am not comfortable making that
change in this CL.

Bug: 142536902
Test: atest SystemUITests && manual
Change-Id: Ie492141005bfd2e24ce0df0665f9585e27b9e0f1

Former-commit-id: a85704052b7a909f43f9d26cb9ba1ddb3896940e