Check that should resume activity directly while relaunch.
When running ActivityLifecycleSplitScreenTests test
testResumedWhenRecreatedFromInNonFocusedStack, we found that an activity
could have an extra paused lifecycle before resumed, this is because
one of the activity state in this stack is pausing, so the activity
behind it be relaunch to pause state.
To prevent another extra lifecycle, we can check that will the activity
goes to resume directly while relaunch.

Bug: 127741025
Test: run tests in android.server.am.lifecycle
Change-Id: I7f0520ec57250f0161741704d0f6ae79ced0139d

Former-commit-id: b5d8f8b185c69f9d02835a75bbae2ecf9363a227