Make the default user the current user (not 0)
In #clearValuesAndCheck, values are placed in the value map based on a
key that depends on the current user. When these values are checked
in #call, it uses default user 0 (unless CALL_METHOD_USER_KEY is
populated).

Considering we don't even use getStringForUser when checking the values,
it seems to make sense to check against the userId used to set the
values in #clearValuesAndCheck.

Test: atest TestableSettingsProviderTest
Change-Id: I0ee9313b1394515d9405344482dc1779d15c0819

Former-commit-id: ed590eec31c5f96b0e41747e3608859bc0e1e5f0