Add explicit private audio recording request
Add the possibility for apps to indicate that their
capture use case is private and that a privileged Assistant
should not be able to capture concurrently.
This allows to override the default behavior tied to the audio
source (e.g VOICE_COMMUNICATION is private by default but
UNPROCESSED is not).

- Add setPrivacySensitive() API to AudioRecord.Builder and
isPrivacySensitive() to AudioRecord
- Add setPrivacySensitive() and isPrivacySensitive() APIs to MediaRecorder.

Bug: 137850106
Test: CTS AudioRecordTest and MediaRecorderTest, manual audio smoke tests

Change-Id: I982b17cf08a9e5e0cf0501117bcaa6930b291f94

Former-commit-id: 597f0b5e95c24860e9492f95796893fb219b2a67