Revert "Revert "Move DozeServiceHost out of StatusBar.""
This restores the refactoring of DozeServiceHost.

It was originally reverted because of strange behavior seen on the
lock screen that prevents the phone from being unlocked. Upon
review, I found that a call to StatusBar.updateDozingState() had been
erroneously omitted from StatusBar.onDozingChanged. Restoring this
call seems to make the observed problem go away.

Bug: 143484506
Test: atest SystemUITests
Change-Id: Ic024db5a3f2bdca3e877aa067f21cab6a5a45454

Former-commit-id: 1193aa49d793482b5d6bb14d09e5958f56eecbdd