Reland: Rework platform version to hide codenames.
The public platform version no longer can be a codename, it is
always the most recently released platform.  A new build property
and API provides either the offical version or the current codename
as appropriate.  This will avoid breaking apps that look at the
platform version while development is under a codename.

Bug:  143175463
Test: manual
Exempt-From-Owner-Approval: Relanding change with zero delta in this project.
Change-Id: I4407df40151d884293fa9d301a401b65bedb5527

Former-commit-id: db0074542c2899fd9a5856d9e9e8e2d11bbe4438