Remove rollbacks with invalid sessions
This could happen to sessions with rollback enabled that are abandoned
before reboot. We need to remove them from mRollbacks so we won't do
anything funny when user is unlocked to cause crash.

See b/143516169 #1 for a detailed explanation for how crash happens.

Bug: 143516169
Test: atest StagedRollbackTest#testPreviouslyAbandonedRollbacks
and search device logs for "FATAL EXCEPTION IN SYSTEM PROCESS" to make
sure system_server doesn't crash inside RollbackManager.

Change-Id: I2c8371be6a9d0e90459ee8f4695f14a94b3aeaca

Former-commit-id: a4ce433eae59ce59850e378125c28882059e458b