DO NOT MERGE - Removed FLAG_EPHEMERAL from pre-created users.
When a User is pre-created with that flag, its CE & DE keys are not persisted on storage. Hence,
when that pre-creatd user is converted to full-user, the expected keys don't exist and the user
doesn't boot properly.

This change fix the issue by making the pre-created user non-ephemeral, and changing it to
ephemeral when it's converted to a "real" user.

Test: manual verification on Automotive by creating a new user than deleting it using the UI -
that creates and switches to a new guest user (which was pre-created); without this change,
that user  would be hanging with a "Android is starting..." message

Bug: 143092698
Fixes: 143553590

Change-Id: If595b29356b5b50c812f2888cda6e636f18a0d4b

Former-commit-id: b286462f6859ee4fd5f890e37cd614361d1696b1