Regenerate API signature files to correct annotation order
This does not change the API in anyway other than to enforce a
consistent, sorted order for all annotations; sorted by qualified
name of the annotation.

A simple bug in metalava meant that annotations were not sorted on an
API item that had exactly two annotations in the source. Fixing the
bug forced the regeneration of the signature files.

Bug: 143556317
Test: m checkbuild
Change-Id: I0a3d967571bb2f796f65c96a9ce9121e3538bad9

Former-commit-id: 7ba79915689be6165ab444122c97721cb5feefdd