Update callbacks when media session has changed
In some cases it is possible to destroy the media session but still have
a notification in the shade. An app can then make a new session and
update the same notification with it.

Since we remove the callback/listener when a session is destroyed, we also
need to add them again if the session has changed.

Fixes: 142909875
Test: manual; atest
com.android.systemui.statusbar.notification.row.wrapper.NotificationMediaTemplateViewWrapperTest

Change-Id: Iefaf28b931f5438be005455af1e0fbb40fed2d60

Former-commit-id: d1eb0a8c37b1dfa76674d53270e6e64bc71f7a5d