Remove libcore implementation source from hiddenapi-mappings
There is an overlap between the source code provided by the
metalava-api-stubs-default and that provided by the
non_openjdk_java_files and openjdk_java_files filegroups. That causes
problems when switching the API generation from libcore implementation
code to libcore stubs code.

Removing the libcore implementation sources will remove any mappings
from class/member signatures to source location for any files from
libcore but that should not be a problem as that was only used for
batch addition of UnsupportedAppUsage annotations and that has alreadu
been done.

Bug: 143864733
Test: m hiddenapi-mappings && m checkbuild
Change-Id: Id47dcb8fc497c9bafa6f501cb2debb835a5f6942

Former-commit-id: 1e0342fd2ca2785a2edfbc26108a60101aede2e1