Make CommandQueue optional.
This CL reorgnized providers of StatusBar related classes into
StatusBarModule and StatusBarDependenciesModule. The latter only
contains a provider for CommandQueue, but other common StatusBar
related classes can be added to it in the future. In this way,
changes to CommandQueue construction (or related classes) do
not need to be repeated across multiple projects.

Bug: 143844964
Test: atest SystemUITests
Change-Id: If7e5fcf17eb4c3fee35ac48cc3565d650a916d6b

Former-commit-id: e9513350963ea4190fe52243be6d0219388fdcb2