vibrator: handle aidl HAL dying
Vibrator just tries to reconnect when commands fail, but because of the
HIDL->AIDL shim's handling of errors, we were dropping binder-level
errors, and the client code wasn't properly handling them. Now, we
convert dead object and other errors appropriately so that the client
code is given the errors it is expecting.

Bug: 141828236
Test: kill aidl vibrator HAL repeatedly, and ensure that it reconnects
    and starts handling vibrator commands.

Change-Id: I9719ec030f514905f1c5e57a1bdf455956d88e76

Former-commit-id: ee3a589204a5f621f0a0f8eac81f074f25724a8f