Ensure that maps are only accessed in BG thread
If the maps are accessed from main thread (as well as bg thread), they
could be in an inconsistent state. In particular they could have nulls
for non-nullable types.

This CL asserts that createFilter is only called in BG thread. It also
makes sure that only non-nulls are passed to IntentFilter#addAction.

Test: atest UserBroadcastDispatcherTest
Fixes: 144039089
Change-Id: I58e66bc9a861c70c830fcbd4476d6dfc991c4ad0

Former-commit-id: 213049e66f5a1c6d88fba82d7fda415c2776a9ea