Add tests to WallpaperBackupAgent
We're currently seeing the WallpaperBackupAgent generate data
for a backup even when nothing has changed. In order to verify
that any fix resolves this we should put appropriate tests in
place which first verify the issue and then break and need
updating when the issue is resolved.

This CL adds some initial tests to give us the test infrastructure
we can use to create the null-diff tests.

Test: atest WallpaperBackupAgentTests
Bug: 140995339
Change-Id: I188e39f9a032b0b9be01faa77597a8698d9e4c57

Former-commit-id: 196d27529251d79a13098cb390a3062006add11c