MediaRouter: send control hints from provider to client
This CL enables providers to send control hints to the clients
so connections can be established between clients and providers.

This CL also adds a hidden flag such that MR2 can notice whether
the route is selected by itself or MR2Manager.

Uses of Mockito is completely removed from the test for better test
environment.

Callbacks will be unregistered automatically when a test is
ended so there will be no unregistered callbacks remaining after
test failures.

Test: atest mediaroutertest
Change-Id: I1411bad3ddd7890fbf76020107f870daa2413077

Former-commit-id: f7d5e044a8ca2602d996dde9f4fd66805f3aca86