remove framework_base_subdirs
Metalava no longer requires it at least for droidstubs in this
directory. Even when local_sourcepath is not set, the -sourcepath option
is set to the module directory (frameworks/base in this case), and
metalava tries to find package.html files under the directory to know if
a package is hidden or not.

Bug: 144264106
Test: m

Change-Id: Idfa2aa392ae1da9582e7691a8c06c986baad5088

Former-commit-id: e53786e864fd3f3097603ef9dfcfcb6b604ee7b1