Save pinned stack bounds when there is no reparenting
This is regression caused by ag/9577803 that we no longer do task
reparenting when exit PiP, in which we signal the SysUI to save reentry
fraction and bounds for PiP window.

Fixed this by moving the save logic to ActivityStack#dismissPip when the
PiP window is about to be transformed to fullscreen.

Bug: 144368633
Test: atest PinnedStackTests#testEnterPictureInPictureSavePosition
Change-Id: I895c33d34b268aed3f8adbf4c905837775b8e30c

Former-commit-id: 7d0c2c0d14ff3317246b8a0d97cbedd1b9c0e77e