Pipe featureId from app to noteOp in BT code
FeatureIds are not yet available, hence in BTManager we assume always a
"null" featureId.

The effect of this change is that for apps that opt into using
featureIds, they will have one BluetoothAdapter per feature, not one per
process as before. In my testing this caused no problem. Most apps
won't use featureIds, hence for most apps there is no change in
behavior.

Test: used bluetooth
Bug: 136595429
Change-Id: Ic40326ea331c60f764f213bb2673cb4c49a81604

Former-commit-id: 22209940e8d0847d052cd5b5fbc8e4ba62fcb4e4