Fix animation when launching AppInfo in LiveTile task
When enabling LiveTile, in overview screen, if launching
AppInfo activity from LiveTile task, since the task view is not
screenshot but is a reparented task surface by RecentsAnimation,
so the wallpaper target will be recents,

But because we ignored to upgrade wallpaper transition because the
topClosingApp is not wallpaper target, it will cause launching AppInfo
activity with TRANSIT_TASK_OPEN or TRANSIT_TASK_TO_FRONT instead of
TRANSIT_WALLPAPER_CLOSE transition.

The fix is to add a check in
RecentsAnimationController#isWallpaperVisible to check if the live-tile
task is in recents animation task and currently wallpaper is visible and
live-tile is on-top of the recents.

So that the topClocingApp and wallpaper target will be the live-tile task and
can upgrade to wallpaper transition.

Fix: 143775967
Test: manual in test steps:
1) long press launcher to launch menu -> Home setttings
2) Enable quick step live-tile.
3) Launch youtube
4) Swipe-up from navbar to launch recents and make sure yotube is still
   playing video.
5) Press youtube itcon on the task view and choose launch AppInfo
6) See if the animation is scaled-up from recents.
   (i.e.: TRANSIT_WALLPAPER_CLOSE transition)

Change-Id: I3071ba7410f0c1b7ad98450a3d172e22c695c015

Former-commit-id: ae078778c8ed07f0489d18d937a8837593589b43