ImageDecoder: compare Strings with .equals()
Bug: 143231863
Test: Infeasible?

Previously we compared with "==", which requires that the two String
objects be the same to return true. We want to return true if the two
Strings are logically equal, even if they are different objects.

Prior to this commit, we may have accidentally called
openAssetFileDescriptor when we should have called
openTypedAssetFileDescriptor.

Change-Id: I9229039b752bafa9a9b85b914a62093dff1eec34

Former-commit-id: 637efbaa0a75c72461b325f3648ab4dababad0bb