Expose setFixedToUserRotation via IWindowManager.
With the parent CL that allows the topmost fullscreen activity to set
the display rotation when freeform windows are visible, OEMs may need
this method to easily regulate the display rotation behavior on external
displays, or primary displays when a keyboard is attached. Therefore
let's expose it to for SysUI to use it.

Also change its name for consistency, especially to differentiate this
and mIgnoreRotateForApps in DisplayContent.

Bug: 144185986
Test: Command line can still set it.
Change-Id: Id1233869c30b66bde3550267dbb21a84076c223d

Former-commit-id: a3f19030ed06cba1b3fbbb4af72b4b330fa9c91a