MediaRouter: Allow empty name for routes
I thought that routes with empty name are invalid and providers
shouldn't publish those.

BT devices seem to be able to have empty name, however, so we have
to choose between
 a) ignoring empty named bluetooth routes
 b) allowing empty named routes

If a BT device was invalid, BT or Audio should hide the device.
So even the name is empty, the route may be valid and we choose b).

Test: atest mediaroutertest
   && atest A2dpServiceTest

Change-Id: I9b4866711bc4f4c2c13e790519ea1f4d408fa5d1

Former-commit-id: d3e53c5ec3b4076324382be4654f82bb35c9f479