Allow registration of observers without specifying packages
This is a prerequisite for integrating RescueParty into PackageWatchdog.
Currently, an observer may only be registered for an existing
ObserverInternal, and the only way to add an ObserverInternal to
mAllObservers is to supply a set of packages.

Observers are pruned if they do not contain any packages. Change this
by adding a method to the PackageHealthObserver interface to indicate
whether an observer may process failures without an explicit set of
packages to monitor.

Test: atest PackageWatchdogTest
Bug: 136135457
Change-Id: Ic97f5d9892af949fe5bbc71ecf4b429f569964e9

Former-commit-id: 84e99020aa7be6067ded85c979bad91dde77e340