Rollback everything that is available when crash is unattributable
When there is an unattributatble crash in native process, there is no
way to tell which staged install is the culprit. As such, the safest
approach is to rollback everything that is available.

Notes:
- Logging events require moduleMetadataPackage. They are only available
when we are rolling back "modules". We don't log when the rollback is
not related to "modules".

Bug: 141843321
Test: atest StagedRollbackTest#testNativeWatchdogTriggersRollback
Change-Id: Ib96dcd8c728a3b5a22cce672128e475c5ad4ac8c

Former-commit-id: 7bd8788c978654a8161777469bf341dd18932198