Remove call to setLayerStack for BlackFrame Surface
There's no need for the black frame to call setLayerStack since it's a
child of the overlay layer. Calling setLayerStack does nothing, but it
will produce an error message in SurfaceFlinger since the caller is
trying to set a specified layerStack on a layer that's not the root.
Deleting thihs code to remove the error spam in logcat.

Also removed it from the Watermark code since it's also a child of the
overlay layer.

Bug: 144824244
Test: No more error log about invalid layerStack in SF
Change-Id: I2f64d50d81af9caf8debe289c4a8eda7c2a61031

Former-commit-id: deb69378a55c665aa0ddd09d71e7d776fdf7e9d6