Allow UserSystemPackageInstaller to ignore OTAs
This cl provides a temporary workaround for a bug.

During OTAs, the UserSystemPackageInstaller goes through all the system
packages and installs them on a user if the package is whitelisted (or
if the feature is no longer being enforced). This ensures that if a package
becomes whitelisted during an OTA, it will be installed on the user,
even if that package had not been whitelisted at the time the user was
created.

This is currently causing ManagedProvisioning to accidently allow
non-required system packages to be installed during OTAs. We
therefore add a mode to this feature to allow ignoring OTAs, thereby
temporarily avoiding this bug.

Bug: 143200798 (temporary workaround)
Test: atest UserSystemPackageInstallerTest

Change-Id: I0a8a44f89c13f5772df57a6cfeea6f484d53d9ae

Former-commit-id: 0274a6ed570a089739858977e5930f16e1dc0e76