Clarify the term in accessibility string
The string is to describe the window when the caption get selected. It
can be a freeform window, or a fullscreen window in freeform display.
Instead of describing the window state, we may want to directly describe
the window itself. Because the caption bar is selected, describe the
caption may be a better choice.

Bug: 144963245
Test: Enable accessibility TalkBack and touch on the caption, to hear
      the speech.

Change-Id: I9b3d570acb1666c6803449088d0b2f8dee370590

Former-commit-id: a682a89a7649817f6d2fe3b4db4eb36e6fd60247