Fix issue where lock wallpaper settings was not being remembered.
If there was a lock wallpaper, and the user then sets the lock
screen to the same as the system wallpaper, the "-1" value for the
lock wallpaper ID (which represents 'use the same as the system
wallpaper') was not remembered. This caused a backup per-day to be
made for the wallpaper.

Fixes: 140995339
Test: atest WallpaperBackupAgentTests
Change-Id: Ia171d21e8c12755bdae418a52ca854a4bca3a695

Former-commit-id: 05b5c6f2fbf56834106b02a66afd04bd717501fb