Prevent Insets side visibility change during animation
If app requests to show IME as soon as app resumes, the animation is
prepared and IME window is set visibile/invisible at start of the
animation. Also, control is set to the window when it receives focus.
This results in control showing the IME before even animation starts
creating a flicker.
Solution is to not applyVisibility when control is set if IME (or
another side) is already animating.

Fix: 144503524
Test: Manually using steps in bug.
      atest InsetsControllerTest InsetsSourceConsumerTest

Change-Id: I5271d906772510cd819f75dc8eed141f4248b34f

Former-commit-id: 93ea15a536e19d6e72a82ca0edb87f404650a3f8