Disable resetting OP_LEGACY_STORAGE
When app preferences are reset AppOpsService toggles the appops for
all apps. This gives the Op watchers a chance to reset any state.

For OP_LEGACY_STORAGE case, the StorageManagerService listens for
changes so it can remount the appropriate storage view for the app.
This remount operation is expensive and unnecessary when resetting
app preferences because the view would not change at the end of the
reset.

Test: No remount uid logs from op changes when resetting app preferences
Bug: 141678467
Change-Id: I86c36786652b36504e624e3f021ad73396328c5d

Former-commit-id: c72c9d17c54346157604beb8f049d581bd33d721