Add a compat change to opt-in to latest SELinux domain.
Refactor getSeInfo into a method, if the app doesn't share a
sharedUserId, and the change is enabled use targetSdkVersion R (1000)
for the seInfo domain.

Add a listener to the compat change that checks if seInfo has changed
(as a result of enabling/disabling the change), and if it did updates it,
and calls prepareAppDataAfterInstallLIF to make sure the data is
correctly labeled.
The listener will be called when an override is introduced or removed
for the change.

Bug: 143539591
Test: atest com.android.server.pm.SELinuxMMAXTest
Test: Install facebook (target SDK 28), login +download a photo, turn on
      the change, open facebook, download another photo, disable change,
      open facebook again, download another photo. Everything works as
      expected, photos are accessible, profile is setup.

Change-Id: If5ecf490e781db13d84656b36d0740a00ee124b5

Former-commit-id: c3c2e4de2d24e17cc05482abccb2d63abd811d75