Bind/unbind notif children content dynamically
Unbind notificiation children content when they won't be visible anyway
in order to save memory. Bind again lazily if we ever do actually need
it.

This is going into the old pipeline so we can start testing it out
early, but ultimately, we also plan to put this in the new pipeline
(though in a form where we try to unbind the row entirely).

Bug: 145748993
Test: Post group on Notify APK. Clear notifs and confirm no weird UX
Test: atest DynamicChildBindControllerTest
Change-Id: I41959627a55fa7041629ce52af5539b3f0c3cecc

Former-commit-id: 3cb379c551a20c0b1c0036b38feb59329d365499