Support freeing contracted/expanded view
There are several places in NotificationContentView (essentially a
FrameLayout holding content views) where we dance around the possibility
of not having a contracted/expanded view and several places where we
assume contracted is the default. We now formalize the expectation in
the view that there may be no content view and allow the API to free the
contracted/expanded view.

Bug: 145748993
Test: locally free contracted/expanded view and smoke test
Change-Id: Iba31677243a3c1cc26dd3d534562f780204c5986

Former-commit-id: f749a1ac7a397390aaac65b2d692a1f87a3bbf46