Throttle stateDescription changed events
Similar to scrolling, state description changes can happen very
frequently, so let's not overwhelm the service when the user is focused
on something like a slider of progress bar.
Isolating and reusing scroll event throttling code to do this.

Test: Added CTS tests in other CL, tried scroll viewws around the
system, and progress bar and slider in talkback test app (which have
already adopted state API).
Fix: 145626718

Change-Id: I4f7d2e526d7d453485615c089b04a8a698828279

Former-commit-id: 35b6f544632ccb39b96a105c2470213237da64b4