Move SystemUI to /system_ext
we plan to forbid hidden API in product partition.
So this app should be moved to /system_ext because it uses hidden API.
It is not permanent, so they can be moved back to /product due to
needs if there isn't hidden usage anymore.

Bug: 137908189
Test: m
Test: check whether device boots successfully
Test: atest SystemUITests
Exempt-From-Owner-Approval: cherry-pick from internal
Change-Id: I54f86496d4a0045af3576091cce00e5cd285b4f8
Merged-In: I54f86496d4a0045af3576091cce00e5cd285b4f8

Former-commit-id: e4ac0ac9ad2c29cd84ec9141a130aafb41fba23a