Use root Activity to determine size compatibility in freeform mode.
If a non-resizable activity is over a resizable activity in the same freeform task,
the second activity is going to use size compat mode and it is going to be scaled down
in freeform window.

The goal is to use root activity's properties for the whole task in freeform mode.

Test: manual test steps
    1. Start resizable activity in freeform mode.
    2. Start second non-resizable and fixed orientation activity in same task.
    3. Observe that freeform should not be size compat mode in freeform window.
      atest WmTests:SizeCompatTests#testShouldUseSizeCompatModeOnResizableTask

Bug: 146581799
Change-Id: Ic47db90f4203bc8a5795accb12c3009cf9f9525a
(cherry picked from commit 89b8a17690f65ccc7fe0beac22d0f7236bb2cbbe)

Former-commit-id: aeb50b9a721935936eee9f2cc52f76b87dab2c6e