Add android.annotation.Hide
This change adds android.annotation.Hide class which can be used
to hide an API without using the @hide javadoc tag. This opens up the
possibility of running metalava over the compiled jar, which is expected
to be much faster than when ran with source files.

In the short-term, this will be only used inside services.jar and not for
framework.jar.

Bug: 139391334
Test: m
Change-Id: I589f1deb63e621a96a76d2f613d283c2d3cdfd17

Former-commit-id: 404b9b2cf2dfe15ee1b9e4bd16ade7362d7cd648