Fix suppressing dot from BubbleMetadata suppress dot
If the app has set #suppressNotification on the bubble, then we shouldn't
show the flyout and we shouldn't show the dot on the bubble.

Right now this wasn't working. This CL adds that check to showDot and
showFlyout and adds a test for it.

There was a test in BubbleDataTest that had to be reworked -- it was using
suppressNotification to act like a shade dismissal which isn't quite right
so this CL fixes that as well.

Test: atest BubbleControllerTest BubbleDataTest
Fixes: 145918462
Change-Id: I3d2c40cf0f4d3eb71b5ed10a480c7b4490649a15

Former-commit-id: b25f906535c3878d530b47fd8350c8921c0e3cb6