Remove hidden API usage PhoneConstants.SUBSCRIPTION_KEY/SLOT_KEY
The two constants are hidden APIs and used as extra keys for a few
legacy broadcasts. No new external users are expected. So making a local
copy of them to keep backward compatibility, until the broadcasts are
removed in the future.

Also move ACTION_SUBSCRIPTION_PHONE_STATE_CHANGED constant out of
PhoneConstants since it's only used in TelephonyRegistry for backward
compatibility.

Bug: 140908357
Test: make
Change-Id: Ic7f15851377dcba5353dc6dade61b1dec96bcea5

Former-commit-id: 4c18a3a5a68e0cae2c0755fdf20a7ae6c3017d73