Put userId at the end of parcel
userId is a new field added in ag/8350816, and it is written in between
some old fields when we parcelize it.

That breaks AiAi because some of its code rely on the implementation
details of SelectionEvent.writeToParcel.

In the long run, we may want to update AiAi such that it works
without this assumption, but before that, let's fix that in platform
for now.

Bug: 145962464
Bug: 123232892

Test: atest FrameworksCoreTests:android.view.textclassifier.SelectionEventTest

Merged-In: I98137a095cddf0fb4c600e1e09eef507bb750528
Change-Id: I66f65ac4c81ca964f710404fd081496355276435

Former-commit-id: 642c22b223e40459d74332b629692a94de99ec9a