Fix unexpected process restart of size compat mode
Since hierarchy merged, the difference between full configuration
and resolved override configuration is no longer the same as the
changes of override configuration. It should be more reliable to
explicitly get the change of the requested override configuration
in display.

Also ensure the popup hint window uses the same window type as
its anchor window, so it won't be occluded by activity.

Bug: 139656157
Test: SizeCompatTests#testResetNonVisibleActivity
Change-Id: I146ca1f55227012da58f5e4d6d85e42a41d1f7cb

Former-commit-id: 7f704b5248b749927f49b0cf21425428fbb3b156