Cleans up user info creation in UserGridRecyclerView.
UserGridRecyclerView shouldn't be creating UserInfo objects, especially not fake
ones that only serve as placeholder for a String. Guest and Add User buttons in
the user switcher don't correspond to real users, and as such, shouldn't be tied
to UserInfo objects.

This can also lead to bugs because consumers of UserRecord assume UserInfo
corresponds to a real user.

Instead we pass null in these cases, and make UserInfo nullable.

Bug: 144993580
Test: manual
Change-Id: I1c772564c8e9ae65699943f6a2f8c1d2c75f013b

Former-commit-id: e44dde31c42fa7b5c75464c5ebb4d8b43113c536