Remove ActivityStack#moveHomeStackToFrontIfNeeded
The method seems no longer necessary. The purpose of the original CL 04a0ea60
was to keep the home stack on top while allowing empty stacks existed, which
not making sense now.

And this additional code flow caused exceptions due to unexpected detached
stack access while updating focus in the middle of reparenting a task in
split-screen.


Bug: 145990447
Bug: 145972535
Test: atest ActivityLifecycleSplitScreenTests
Change-Id: I46b8b2044c851edb4e68616c8b7d38115af000e4

Former-commit-id: 074cdf2dbfcab6df898266f0d8c37ed5feb43340