Change touchable region to full screen when modal for non app is set.
When the modal flag is set, use the full screen for touchable region and
remove the modal flag so it can be properly handled when cropped. This
should behave the same when there is no crop since the window will have
the full screen touchable region. For cases with cloning, the cloned layers
will get cropped to the root of the clones so it will properly crop the touch
regions for modal windows.

Test: go/wm-smoke
Bug: 140756730
Change-Id: I5ee151d7500ad3557782cc90db7d8435389eab40

Former-commit-id: f5827d1214cd7c7ef73728687778a1adf30140ff