Run PackageManagerShellCommandTest in Treehugger.
Atest was showing a warning:

WARNING: PackageManagerShellCommandTest is not a valid build target
and may not be discoverable by TreeHugger. If you want to specify a
class or test-package, please set 'name' to the test module and use
'options' to specify the right tests via 'include-filter'.

This change fixes that warning; it may even allow the tests to be run
in Treehugger presubmit.

Test: atest PackageManagerShellCommandTest
Test: atest -b -p services/core/java/com/android/server/pm (tests build, no warning)
Change-Id: I6f507ec005fc2e946358e2d54ed4d255705c5e9b

Former-commit-id: 5995bc587b0222c5ae37ba68b879782c7d494394