Fix summary cancel logic for experimental bubbles
Some apps seems to cancel their summary notification when you open a chat
and interact with it. For experimental bubbles we *don't* remove bubbles
for a cancel because the app doesn't know about them / hasn't adjusted
their notification lifecycle to account for them. I missed the notification
summary path initially, this fixes that.

Test: 1) go/use-bubbles : have experiment running
      2) get a bunch of notifs from app so there's a summary
      3) make one of the children bubble via longpress menu
      4) expand bubble & cause summary to be canceled by the app somehow
      => Note that the bubble remains on screen & is interactive
Bug: 145763712
Bug: 146093179

Change-Id: Idc27a99c87b41a47894a9460bdc56e2389afda28

Former-commit-id: 840942243fc0cff8120df08ff98dec18a63e8e46