Deprecate some UICC related APIs in TelephonyManager
In specific, the following APIs are marked deprecated:
- TelephonyManager#iccOpenLogicalChannel
- TelephonyManager#iccCloseLogicalChannel
- TelephonyManager#iccCloseLogicalChannelBySlot
- TelephonyManager#iccTransmitApduLogicalChannel
- TelephonyManager#iccTransmitApduLogicalChannelBySlot
- TelephonyManager#iccTransmitApduBasicChannel
- TelephonyManager#iccTransmitApduBasicChannelBySlot
- TelephonyManager#iccExchangeSimIO
- TelephonyManager#sendEnvelopeWithStatus

The comments of the APIs are updated, suggesting to switch to
android.se.omapi.SEService APIs instead.

Telephony APIs do not provide any resource management for opened
logical channels, so it's easy for apps to open channels and
never close them. Secure Element Service does provide resource
management, so we are directing apps to use those APIs instead.

Bug: 146360599
Test: N/A
Change-Id: Id44402c637aaa77c77ee7672e0ceddf1bee20f7c

Former-commit-id: d9832ed2ed78f0ac39a57192703d9ee6a9092c74