wifi: add interface name to ScanResult
Add interface name to ScanResult to allow connection on a secondary
interface.
In the initial support for multiple chips, a single
ClientModeImpl supports connection to multiple interfaces,
but only one at a time.
When a new connection is needed, the ClientModeImpl is
handed a WifiConfiguration with attached ScanResult to
describe the connection. The ClientModeImpl needs to
figure out on which interface to perform the connection.
Having interface name as part of ScanResult makes it very
easy to figure out the interface. It is also very easy to
add as the scanner implementation always have this information
at hand.
Without the interface name, there is not enough information
to figure out the exact interface (chip) to which the
connection belongs.

Bug: 147495504
Test: Manual - enable wifi and connect to an access point
Test: atest com.android.server.wifi
Test: atest FrameworksWifiApiTests
Change-Id: I9e2f102f27eab01dbc407ea9cbb0bd9d6819395a

Former-commit-id: 0234e4c4c926043e2c9e6c8f71407f1e2ff03c6a