Introduce TvSystemUIRootComponent
Introduces TvSystemUIFactory and TvSystemUIRootComponent. The purpose of
the former is to substitute "default" SystemUIRootComponent for an
instance of the latter. This CL keeps TvSystemUIRootComponent completely
identical to SystemUIRootComponent, since the goal here is to make the
split. The following CL(s) would modify TvSystemUIRootComponent so that
it only binds SystemUI components that are relevant on TV.

Change-Id: Ib7c6dce2ab65373a1cb25f41fdfcf69fad24ea75
Bug: 146188087
Test: make SystemUI, atest SystemUITests

Former-commit-id: 7842bac5e48a2c9659b44b247add52f5b4e61c74