Add API stubs for module APIs
The following two stub libraries are added:

- android_module_app_stubs_current: @SystemApis that are available for
modules implemented as apps. @SystemApis with client=PRIVILEGED_APPS
and client=MODULE_APPS are included here.

- android_module_lib_stubs_current: @SystemApis that are available for
modules implemented as bootclasspath jars. @SystemApis with
client=PRIVILEGED_APPS, MODULE_APPS, and MODULE_LIBRARIES are all
included here.

For now, the compatibility check against the last release API is not
done, but this is not a problem because this is the first time these
APIs are released.

Bug: 139391334
Bug: 140202860
Bug: 146727827
Test: m

Change-Id: I0254a5f7751c58029fa7a8d5c342b731a7deb69a

Former-commit-id: e309516406a1ec2ce095713cc3adfacdedbcbefe